# perf(core): skip checksum for cache entries when it isn't required